Fix router.currentRoute from not being accurate if user hits back button
Just setting router.currentRoute when we call router.goTo isn’t good
enough because the view changes when the user hits the back button, but
Apple TV does that automatically and without notifying us.

So we set the routing context onto the document when we present it, and
we can look at the last item in navigationDocument to determine the
current document and the correct current path.